requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cloudinary==1.41.0
cryptography==44.0.0
Deprecated==1.2.15
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
h11==0.14.0
idna==3.10
Jinja2==3.1.5
limits==3.14.1
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.0
